corrected tables exercise

+++ albums_migration.sql
@@ -5,7 +5,7 @@ CREATE TABLE albums (
 	artist VARCHAR(100) NOT NULL,
 	record_name VARCHAR(100) NOT NULL,
 	release_date YEAR(4) NOT NULL,
-	sales_in_millions FLOAT NOT NULL,
+	sales_in_millions FLOAT (15,2) NOT NULL,
 	genre VARCHAR(100) NOT NULL,
 	PRIMARY KEY (id)
 )
